requirements.txt
absl-py==0.1.10
bleach==1.5.0
captcha==0.2.4
enum34==1.1.6
html5lib==0.9999999
Keras==2.1.3
Markdown==2.6.11
numpy==1.14.0
Pillow==5.0.0
protobuf==3.5.1
PyYAML==3.12
scipy==1.0.0
six==1.11.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
virtualenv==15.1.0
Werkzeug==0.14.1
